+++ source/Energinet.DataHub.MeteringPoints.EntryPoints.Outbox/RequestHandlers/CreateMeteringPointEventHandler.cs
@@ -14,7 +14,8 @@
 
 using System.Threading;
 using System.Threading.Tasks;
-using Energinet.DataHub.MeteringPoints.Infrastructure.IntegrationServices.Dispatchers;
+using Energinet.DataHub.MeteringPoints.Infrastructure.Integration.Dispatchers;
+using Energinet.DataHub.MeteringPoints.Infrastructure.Integration.Messages;
 using MediatR;
 
 namespace Energinet.DataHub.MeteringPoints.EntryPoints.Outbox.RequestHandlers
