+++ source/Energinet.DataHub.MeteringPoints.Infrastructure/Energinet.DataHub.MeteringPoints.Infrastructure.csproj
@@ -67,8 +67,4 @@ limitations under the License.
       <ProjectReference Include="..\Energinet.DataHub.MeteringPoints.Application\Energinet.DataHub.MeteringPoints.Application.csproj" />
     </ItemGroup>
 
-    <ItemGroup>
-      <Folder Include="IntegrationServices\Handlers" />
-    </ItemGroup>
-
 </Project>
