+++ source/Energinet.DataHub.MeteringPoints.Infrastructure/Integration/Dispatchers/IntegrationEventToEventHubDispatcher.cs
@@ -12,21 +12,16 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using System;
-using Energinet.DataHub.MeteringPoints.Infrastructure.IntegrationServices.Dispatchers;
+using Energinet.DataHub.MeteringPoints.Application.Transport;
+using Energinet.DataHub.MeteringPoints.Infrastructure.Integration.Channels;
 
-namespace Energinet.DataHub.MeteringPoints.Infrastructure.IntegrationServices.Helpers
+namespace Energinet.DataHub.MeteringPoints.Infrastructure.Integration.Dispatchers
 {
-    public class IntegrationEventTypeFactory
+    public class IntegrationEventToEventHubDispatcher : MessageDispatcher
     {
-        public static Type GetType(string type)
+        public IntegrationEventToEventHubDispatcher(MessageSerializer serializer, AzureEventHubChannel channel)
+            : base(serializer, channel)
         {
-            if (typeof(CreateMeteringPointEventMessage).FullName == type)
-            {
-                return typeof(CreateMeteringPointEventMessage);
-            }
-
-            throw new ArgumentException("Integration Event type is not implemented.");
         }
     }
 }
