+++ source/Energinet.DataHub.MeteringPoints.Infrastructure/Integration/Messages/CreateMeteringPointEventMessage.cs
@@ -14,11 +14,9 @@
 
 using Energinet.DataHub.MeteringPoints.Application.IntegrationEvent;
 using Energinet.DataHub.MeteringPoints.Application.Transport;
-using Energinet.DataHub.MeteringPoints.Domain;
-using Energinet.DataHub.MeteringPoints.Domain.MeteringPoints;
 using MediatR;
 
-namespace Energinet.DataHub.MeteringPoints.Infrastructure.IntegrationServices.Dispatchers
+namespace Energinet.DataHub.MeteringPoints.Infrastructure.Integration.Messages
 {
-    public record CreateMeteringPointEventMessage(string Gsrn, string MpType, string GridAreaId, bool Child) : IOutboundMessage, IRequest, IIntegrationEvent;
+    public record CreateMeteringPointEventMessage(string Gsrn, string MeteringPointType, string GridAreaId, bool Child) : IOutboundMessage, IRequest, IIntegrationEvent;
 }
