+++ source/Energinet.DataHub.MeteringPoints.Infrastructure/Processing/Contracts/Contract.proto
